+++ cotests/bench/__init__.py
@@ -1,2 +1,2 @@
 from .co_case import AbstractCoCase
-from .typ import CoException, AbstractTestCase
+from .typ import CoException
